refactor(ui): introduce typescript
