+++ src/options.html
@@ -15,8 +15,7 @@
     <script src="/vendor/jquery/jquery.address.js"></script>
     <script src="/vendor/semantic/semantic.min.js"></script>
     <script src="/vendor/sortable/Sortable.min.js"></script>
-    <script src="/vendor/vue/vue.runtime.min.js"></script>
     <script src="/vendor/vue/vuedraggable.umd.min.js"></script>
-    <script src="ui/options.js"></script>
+    <script src="ui/options.ts"></script>
   </body>
 </html>
