+++ src/popup.html
@@ -13,8 +13,7 @@
     <script src="/vendor/jquery/jquery.min.js"></script>
     <script src="/vendor/semantic/semantic.min.js"></script>
     <script src="/vendor/sortable/Sortable.min.js"></script>
-    <script src="/vendor/vue/vue.runtime.min.js"></script>
     <script src="/vendor/vue/vuedraggable.umd.min.js"></script>
-    <script src="ui/popup.js"></script>
+    <script src="ui/popup.ts"></script>
   </body>
 </html>
